refactor(cli): extract GitHub PR ops into work_commands_gh.rs — c0bd865c split COMPLETE (phase 4 / c7bdf2df)
The GitHub-side PR operations move to a sibling module:

  Extracted to work_commands_gh.rs:
    - open_pr_and_link — the `gh pr create` → URL parse → PullRequestLinked emit flow
    - gh_default_branch — repo default branch via `gh repo view`
    - extract_pr_number — pure parser for gh's PR-create URL output
    - pr_create_base_branch — hardcoded `rust-rewrite` per card 28f1440c

  work_commands.rs:        1672 → 1517 LOC (-155)
  work_commands_gh.rs:        0 →  ??? LOC (new)

  Original work_commands.rs:                    1823 LOC
  - phase 1 #1042 render:                       -238
  - phase 2 #1046 lifecycle:                    -374
  - phase 3 #1047 git:                          -189
  - phase 4 (this PR) gh:                       -~155
  - phase 5 #1048 review:                       -153
  - FINAL work_commands.rs:                     ~~600 LOC

That's the c0bd865c goal: a sub-thousand-line work_commands.rs containing just the dispatcher / shared types / parser helpers / format helpers. Consumer-embedders (continuum / hermes / openclaw / codex when planning integrations) can read it in one pass; the typed operations live in clearly-named sibling modules.

  - open_pr_and_link (this PR) → git_show_format, git_rev_parse_branch (work_commands.rs, will be phase 3 git module after #1047 lands), auto_spawn_review_card (work_commands.rs, will be phase 5 review module after #1048 lands)

All three are `pub(crate)` so the cross-module path works. When phases 3 + 5 land their respective re-exports route to the right module automatically.

  cargo test work_commands:: → 19 pass (including extract_pr_number_parses_gh_url_or_returns_none, pr_create_base_branch_targets_rust_rewrite which are in the gh-domain test set)
  cargo clippy -p airc-cli --bin airc -- -D warnings → clean
  cargo fmt --check → clean

c0bd865c follow-up final phase. With this PR + the four already in flight, work_commands.rs is split into:
  work_commands.rs          (dispatchers / shared / parsers / tests)
  work_commands_render.rs   (board / roster / manager prints — #1042)
  work_commands_lifecycle.rs (card state machine — #1046)
  work_commands_git.rs      (git + repo-id helpers — #1047)
  work_commands_gh.rs       (GitHub PR ops — this PR)
  work_commands_review.rs   (review card spawn — #1048)

Five focused files, each with a clear responsibility. Embeddable, readable, mergeable into the dec35ec7 / a094aa81 typed-client migration when those land.
