+++ crates/airc-cli/src/main.rs
@@ -78,6 +78,7 @@ mod transport_commands;
 mod update_commands;
 mod work_cli;
 mod work_commands;
+mod work_commands_gh;
 mod work_suggestions;
 mod workspace_cli;
 mod workspace_commands;
