+++ crates/airc-cli/src/work_commands.rs
@@ -558,135 +558,10 @@ pub async fn run_state(
     // auto-spawn review card, board renderers) read one source of
     // truth.
     if card_state == CardState::Review {
-        if let Err(error) = open_pr_and_link(&airc, card_uuid).await {
+        if let Err(error) = crate::work_commands_gh::open_pr_and_link(&airc, card_uuid).await {
             eprintln!("airc: gh pr create skipped — {error}");
         }
     }
-    Ok(())
-}
-
-/// Best-effort: open a GitHub PR for the work card's worktree branch
-/// and link it via `WorkEvent::PullRequestLinked`. Returns Err so
-/// run_state can surface as a warning without aborting the state
-/// transition.
-async fn open_pr_and_link(
-    airc: &airc_lib::Airc,
-    card_id: airc_lib::WorkCardId,
-) -> Result<(), Box<dyn std::error::Error>> {
-    use airc_work::model::{BranchName, PullRequestRef};
-
-    let board = airc.work_board(usize::MAX).await?;
-    let card = board
-        .card(card_id)
-        .ok_or_else(|| format!("card {card_id} not visible in board projection"))?;
-    if card.pull_request.is_some() {
-        // Already linked; nothing to do. Re-running `state review` on
-        // a card that's already been reviewed is a no-op for the PR
-        // side.
-        return Ok(());
-    }
-
-    let short: String = card.card_id.to_string().chars().take(8).collect();
-    let lease_root = lease::lease_root()
-        .ok_or_else(|| "HOME/USERPROFILE not set; cannot resolve ~/.airc/worktrees/".to_string())?;
-    let worktree_path = lease_root.join(&short);
-    if !worktree_path.exists() {
-        return Err(format!(
-            "no worktree at {} — claim was made before card d1b2798d shipped, or with --no-lease-required (re-claim from a worktree to get auto-PR on review)",
-            worktree_path.display()
-        )
-        .into());
-    }
-    let worktree_str = worktree_path.to_string_lossy().to_string();
-
-    // gh pr create — pass --title + --body explicitly from the HEAD
-    // commit's metadata. `--fill` SOUNDS right but its heuristic
-    // sometimes falls back to a slugified branch name as the title
-    // (observed live on cards 53698eb9 and ef168afe — branch
-    // `53698eb9/agents-md-encode-engineering-staff-not-a` became
-    // PR title `53698eb9/agents md encode engineering staff not a`,
-    // which stripped the `docs(agents):` prefix the canary-gate's
-    // bypass regex needs). Reading the commit subject directly is
-    // deterministic and matches what the author actually wrote.
-    //
-    // Card a4fe899f: `gh` does NOT accept `-C` (that's git's flag).
-    // The cwd has to be set via `Command::current_dir(...)` so gh's
-    // own repo-resolution (which scans `git remote get-url origin`
-    // from cwd) picks the worktree's branch.
-    //
-    // Card 28f1440c: `--base` is set explicitly to the workflow's
-    // integration branch (`rust-rewrite`) rather than letting gh
-    // fall back to the repo's GitHub default (`main` here). Main is
-    // the legacy snapshot; every PR landing there bypasses the
-    // substrate work. Hardcoded for the MVP; a follow-up card
-    // surfaces it as configurable for consumers whose integration
-    // branch differs.
-    //
-    // Card 13131f1c: `--title` + `--body` are passed explicitly from
-    // the HEAD commit instead of relying on `--fill`. `--fill`'s
-    // heuristic sometimes substitutes a slugified branch name as the
-    // title (observed on #1032 and #1033 — `docs(agents):`/`fix(ci):`
-    // prefixes stripped, breaking the canary-gate bypass). Reading
-    // the commit subject directly is deterministic.
-    let base_branch = pr_create_base_branch();
-    let subject = git_show_format(&worktree_str, "%s")?;
-    let body = git_show_format(&worktree_str, "%b")?;
-    let create_out = std::process::Command::new("gh")
-        .current_dir(&worktree_str)
-        .args([
-            "pr",
-            "create",
-            "--title",
-            subject.trim(),
-            "--body",
-            body.trim(),
-            "--base",
-            base_branch,
-        ])
-        .output()?;
-    if !create_out.status.success() {
-        return Err(format!(
-            "gh pr create failed: {}",
-            String::from_utf8_lossy(&create_out.stderr).trim()
-        )
-        .into());
-    }
-    let stdout = String::from_utf8(create_out.stdout)?;
-    let pr_url = stdout.trim().lines().last().unwrap_or("").trim();
-    let pr_number = extract_pr_number(pr_url)
-        .ok_or_else(|| format!("could not parse PR number from gh output: {pr_url}"))?;
-
-    // Resolve head from the worktree's git state. Base is the same
-    // value we passed to `gh pr create` above (card 28f1440c) — must
-    // not drift between the substrate's PullRequestRef and the
-    // actual PR target, or downstream consumers (review-spawn,
-    // close-guard) reason about the wrong branch.
-    let head_branch = git_rev_parse_branch(&worktree_str)?;
-
-    let pull_request = PullRequestRef {
-        repo: card.repo.clone(),
-        number: pr_number,
-        head: BranchName::new(head_branch)?,
-        base: BranchName::new(base_branch.to_string())?,
-    };
-    airc.link_card_pull_request(airc_lib::LinkCardPullRequest {
-        card_id,
-        pull_request,
-    })
-    .await?;
-
-    println!("pull_request: {pr_url}");
-
-    // Card ad7e100b Sub-C: with PR linked, spawn a sibling review
-    // card so any peer (other than the author) can claim it and
-    // review the diff. Best-effort and idempotent — a spawn failure
-    // here must not undo the state transition or the PR link, and
-    // re-running `state review` on a card whose review card already
-    // exists is a no-op.
-    if let Err(error) = auto_spawn_review_card(airc, card_id, pr_url).await {
-        eprintln!("airc: review card auto-spawn skipped — {error}");
-    }
-
     Ok(())
 }
 
@@ -714,7 +589,7 @@ async fn open_pr_and_link(
 /// out of the CLI for consumers that bypass it (Continuum, OpenClaw),
 /// but the wire shape is stable — that move doesn't break this
 /// emit.
-async fn auto_spawn_review_card(
+pub(crate) async fn auto_spawn_review_card(
     airc: &airc_lib::Airc,
     parent_id: airc_lib::WorkCardId,
     pr_url: &str,
@@ -754,16 +629,7 @@ async fn auto_spawn_review_card(
     Ok(())
 }
 
-/// Extract the PR number from a gh pr create URL line like
-/// `https://github.com/owner/repo/pull/123`. Returns None for any
-/// shape we don't recognise (so callers can degrade gracefully).
-fn extract_pr_number(url: &str) -> Option<u64> {
-    let trimmed = url.trim();
-    let tail = trimmed.rsplit('/').next()?;
-    tail.parse::<u64>().ok()
-}
-
-fn git_rev_parse_branch(worktree: &str) -> Result<String, Box<dyn std::error::Error>> {
+pub(crate) fn git_rev_parse_branch(worktree: &str) -> Result<String, Box<dyn std::error::Error>> {
     let out = std::process::Command::new("git")
         .args(["-C", worktree, "rev-parse", "--abbrev-ref", "HEAD"])
         .output()?;
@@ -783,7 +649,10 @@ fn git_rev_parse_branch(worktree: &str) -> Result<String, Box<dyn std::error::Er
 /// heuristic sometimes falls back to a slugified branch name (card
 /// 13131f1c). Empty stdout is valid (commits often have an empty
 /// body) and propagates as an empty `String`.
-fn git_show_format(worktree: &str, format: &str) -> Result<String, Box<dyn std::error::Error>> {
+pub(crate) fn git_show_format(
+    worktree: &str,
+    format: &str,
+) -> Result<String, Box<dyn std::error::Error>> {
     let out = std::process::Command::new("git")
         .args([
             "-C",
@@ -802,47 +671,6 @@ fn git_show_format(worktree: &str, format: &str) -> Result<String, Box<dyn std::
         .into());
     }
     Ok(String::from_utf8(out.stdout)?)
-}
-
-/// Card 28f1440c — the integration branch every per-card PR opens
-/// against. Hardcoded to the airc substrate's working branch for
-/// MVP; a follow-up surfaces this as configurable (e.g. via
-/// `.airc/work.toml`) for consumers whose integration branch is
-/// different.
-///
-/// MUST NOT change to a runtime read of `gh repo view --json defaultBranchRef` —
-/// that returns the repo's GitHub `default_branch` (`main` on this
-/// repo today), and every PR landing on `main` bypasses
-/// `rust-rewrite`'s substrate work. The whole point of this card
-/// is to refuse that fallback.
-pub(crate) fn pr_create_base_branch() -> &'static str {
-    "rust-rewrite"
-}
-
-#[allow(dead_code)]
-fn gh_default_branch(worktree: &str) -> Result<String, Box<dyn std::error::Error>> {
-    // Card a4fe899f: `gh` does NOT accept `-C`; set cwd via
-    // `Command::current_dir(...)` so `gh repo view` resolves the
-    // worktree's origin remote, not the shell cwd's.
-    let out = std::process::Command::new("gh")
-        .current_dir(worktree)
-        .args([
-            "repo",
-            "view",
-            "--json",
-            "defaultBranchRef",
-            "--jq",
-            ".defaultBranchRef.name",
-        ])
-        .output()?;
-    if !out.status.success() {
-        return Err(format!(
-            "gh repo view failed: {}",
-            String::from_utf8_lossy(&out.stderr).trim()
-        )
-        .into());
-    }
-    Ok(String::from_utf8(out.stdout)?.trim().to_string())
 }
 
 /// Card a1bc62b3 — gate on direct CLI state writes.
@@ -1618,16 +1446,21 @@ mod tests {
     #[test]
     fn extract_pr_number_parses_gh_url_or_returns_none() {
         assert_eq!(
-            extract_pr_number("https://github.com/CambrianTech/airc/pull/123"),
+            crate::work_commands_gh::extract_pr_number(
+                "https://github.com/CambrianTech/airc/pull/123"
+            ),
             Some(123),
         );
         assert_eq!(
-            extract_pr_number("  https://github.com/owner/repo/pull/1  "),
+            crate::work_commands_gh::extract_pr_number("  https://github.com/owner/repo/pull/1  "),
             Some(1),
         );
         // Non-numeric tail → None (degrade gracefully).
-        assert_eq!(extract_pr_number("https://example.com/no-pr-here"), None);
-        assert_eq!(extract_pr_number(""), None);
+        assert_eq!(
+            crate::work_commands_gh::extract_pr_number("https://example.com/no-pr-here"),
+            None
+        );
+        assert_eq!(crate::work_commands_gh::extract_pr_number(""), None);
     }
 
     #[test]
@@ -1867,7 +1700,7 @@ mod tests {
     #[test]
     fn pr_create_base_branch_targets_rust_rewrite() {
         assert_eq!(
-            pr_create_base_branch(),
+            crate::work_commands_gh::pr_create_base_branch(),
             "rust-rewrite",
             "card 28f1440c: PR target must be the substrate working \
              branch, never the repo's GitHub default ('main' on this \
@@ -1875,6 +1708,6 @@ mod tests {
         );
         // Specifically: it must NOT be 'main'. The doctrine treats
         // main as the legacy snapshot.
-        assert_ne!(pr_create_base_branch(), "main");
+        assert_ne!(crate::work_commands_gh::pr_create_base_branch(), "main");
     }
 }
